Refactoring by review comments.
